Fix gpu reserved memory space to include render buffer

+++ DataMgr/DataMgr.h
@@ -32,7 +32,7 @@ class DataMgr {
           const bool useGpus,
           const int numGpus,
           const int startGpu = 0,
-          const int reservedGpuMem = (1 << 27));
+          const size_t reservedGpuMem = (1 << 27));
   ~DataMgr();
   AbstractBuffer* createChunkBuffer(const ChunkKey& key, const MemoryLevel memoryLevel, const int deviceId = 0);
   AbstractBuffer* getChunkBuffer(const ChunkKey& key,
@@ -66,7 +66,7 @@ class DataMgr {
   std::vector<std::vector<AbstractBufferMgr*>> bufferMgrs_;
   std::string dataDir_;
   bool hasGpus_;
-  int reservedGpuMem_;
+  size_t reservedGpuMem_;
 };
 }  // Data_Namespace
